Add Contracts folder in Domain and moved there all interfaces
