+++ ProjectManager.Domain/Contracts/IDataSeeder.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ProjectManager.Domain.Interfaces
+namespace ProjectManager.Domain.Contracts
 {
     public interface IDataSeeder
     {
