+++ ProjectManager.Domain/Contracts/Repositories/IProjectTaskRepository.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ProjectManager.Domain.Interfaces
+namespace ProjectManager.Domain.Contracts.Repositories
 {
     public interface IProjectTaskRepository
     {
@@ -14,6 +14,8 @@ namespace ProjectManager.Domain.Interfaces
         Task DeleteTask(ProjectTask projectTask);
         Task<IEnumerable<ProjectTask>> GetAllByEncodedName(string projectEncodedName);
         Task<IEnumerable<ProjectTask>> GetUserProjectTasks(string projectEncodedName, string userId);
+        Task<IEnumerable<ProjectTask>> GetAllUserTasks(string userId);
         Task<ProjectTask> GetById(int id);
+        Task Update(ProjectTask projectTask);
     }
 }
