+++ ProjectManager.Domain/Contracts/Repositories/IUserRepository.cs
@@ -6,12 +6,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ProjectManager.Domain.Interfaces
+namespace ProjectManager.Domain.Contracts.Repositories
 {
     public interface IUserRepository
     {
         Task Commit();
-        Task<IEnumerable<User>>GetAllUsers();
+        Task<IEnumerable<User>> GetAllUsers();
         Task<User> GetUserByEmail(string email);
     }
 }
